# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank Cards API
  description: REST API для управления банковскими картами
  contact:
    name: Bank Cards Team
    email: support@bankcards.com
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags:
        - auth-controller
      summary: Регистрация нового пользователя
      description: Создает нового пользователя с ролью USER (по умолчанию) или ADMIN
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
        required: true
      responses:
        '200':
          description: Пользователь успешно зарегистрирован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '400':
          description: Неверные данные регистрации
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - auth-controller
      summary: Вход в систему
      description: Аутентификация пользователя и получение JWT токена
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Успешная аутентификация
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверное имя пользователя или пароль
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: Время возникновения ошибки
          format: date-time
        status:
          type: integer
          description: HTTP статус код
          format: int32
          example: 400
        error:
          type: string
          description: Тип ошибки
          example: Bad Request
        message:
          type: string
          description: Сообщение об ошибке
          example: 'User not found with id: 1'
        path:
          type: string
          description: Путь запроса
          example: /api/users/1
        validationErrors:
          type: object
          additionalProperties:
            type: string
            description: Детальные ошибки валидации (поле -> сообщение)
            example: '{"username":"Username is required","email":"Invalid email format"}'
          description: Детальные ошибки валидации (поле -> сообщение)
          example:
            username: Username is required
            email: Invalid email format
      description: Ответ с информацией об ошибке
...
    AuthRequest:
      required:
        - email
        - password
        - username
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 3
          type: string
          description: Имя пользователя (уникальное)
          example: john_doe
        email:
          type: string
          description: Email пользователя (уникальный)
          example: john@example.com
        password:
          maxLength: 2147483647
          minLength: 6
          type: string
          description: Пароль (минимум 6 символов)
          example: password123
        role:
          type: string
          description: Роль пользователя (по умолчанию USER)
          example: USER
          enum:
            - USER
            - ADMIN
          default: USER
      description: Запрос на регистрацию пользователя
    JwtResponse:
      required:
        - email
        - id
        - role
        - token
        - type
        - username
      type: object
      properties:
        token:
          type: string
          description: JWT токен для доступа к защищенным ресурсам
          example: >-
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ
        type:
          type: string
          description: Тип токена (всегда Bearer)
          example: Bearer
          default: Bearer
        id:
          type: integer
          description: ID пользователя
          format: int64
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: john_doe
        email:
          type: string
          description: Email пользователя
          example: john@example.com
        role:
          type: string
          description: Роль пользователя
          example: USER
          enum:
            - USER
            - ADMIN
      description: Ответ с JWT токеном после успешной аутентификации
    LoginRequest:
      required:
        - password
        - username
      type: object
      properties:
        username:
          type: string
          description: Имя пользователя
          example: john_doe
        password:
          minLength: 6
          type: string
          description: Пароль
          example: password123
      description: Запрос на вход в систему
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT